BDBOT-2: create user

+++ src/main/java/my/notify/bd/dto/User.java
@@ -3,17 +3,17 @@ package my.notify.bd.dto;
 import java.util.Objects;
 
 public class User {
-    private Long id;
+    private Integer id;
     private String name;
     private Integer year;
     private String month;
-    private Long day;
-    private Long age;
+    private Integer day;
+    private Integer age;
 
     public User() {
     }
 
-    public User(Long id, String name, Integer year, String month, Long day, Long age) {
+    public User(Integer id, String name, Integer year, String month, Integer day, Integer age) {
         this.id = id;
         this.name = name;
         this.year = year;
@@ -22,11 +22,11 @@ public class User {
         this.age = age;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
 
@@ -54,19 +54,19 @@ public class User {
         this.month = month;
     }
 
-    public Long getDay() {
+    public Integer getDay() {
         return day;
     }
 
-    public void setDay(Long day) {
+    public void setDay(Integer day) {
         this.day = day;
     }
 
-    public Long getAge() {
+    public Integer getAge() {
         return age;
     }
 
-    public void setAge(Long age) {
+    public void setAge(Integer age) {
         this.age = age;
     }
